// src/components/Portfolio/types.ts
export type TitleProps = {
  fontSize?: string;
  lineHeight?: string;
};

export type ContentContainerProps = { borderImage?: string };

export type TechIconProps = { theme?: string };

export type UnderlineProps = { margin?: string };

export type LinkProps = {
  borderImage?: string;
  hover?: string;
  fontSize?: string;
  margin?: string;
};